requirements.txt
rdflib
xmltodict
lxml
pyyaml
pytz
pymysql
autopep8
pytest
nose
requests
cryptography
coverage
pytest
pytest-cov